requirements.txt
alpaca_trade_api
numpy
sounddevice
pandas
python-dotenv
pandas_ta